Update search utils
